Add param "ignoreEnvironmentKeys"
New param allows ignore list of keys we don't want output to destination .env file as we may only want them declared as environment variables for Lambda function
